Fix SIGSEGV
If a template is mutated so that a subread is no longer aligned to any
part of it, we wind up in big trouble.  In particular, the code tries
to index a matrix at column 0 - 1 = Max(size_t), which quite often
leads to errors.  This problem is pretty deep in the code, and can
really only happen when there is a very noisy subread that does not
belong with the others.

This error can strike when:

	* A template is initially created
	* A mutation is scored
	* A mutation is applied (which triggers a refill of the matrices)

How to handle it?  We originally threw a SmallTemplateException in
these cases, and given that this is a very rare case, it seemed like
the cleanest way to handle this.  However, to properly handle this, I
split up the code so that instead of mutations being applied and then
immediately recalculating the score, a mutation is applied, empty
alignments are removed, and then the score is recalculated.  This logic
should avoid this case, but exceptions are peppered throughout the code
to verify that we never have a template span of size 0 and robustify
against SIGSEGV chances.

This change revealed a number of other issues, in particular:

	* Scoring insertions at the end of templates could lead to a different
SIGSEGV, I have altered that as well.
	* The convention that the intial basepair have a match probability of
1 was not enforced throughout, leading to errors.  I have fixed this.

Additionally, the logic for picking the range of rows to use in
ExtendAlpha was far too complicated and led to errors.  Rather than
micro-guess each column, I use the same range for all columns we are
extending over.