Try out simple "insertion burst" filter
This commit is a collection of quick edits that enable read localized
editing to avoid regions with many insertions affecting consensus.  The
code works by pseudocounting insertions to generate masks that locally
filter noisy regions.  The implementation is only to quickly test the
idea, as this is not production code.